# dashboard/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Flaglane API",
    "description" : "Feature flags and remote configuration. `/api/**` serves the dashboard with a JWT; `/sdk/**` serves SDKs with an API key.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.1.0"
  },
...
  "paths" : {
    "/api/projects/{projectKey}/flags/{flagKey}/config/{envKey}/rules" : {
      "get" : {
        "tags" : [ "Targeting" ],
        "summary" : "Get a flag's rules in an environment",
        "description" : "In priority order.",
        "operationId" : "rules",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The rules, in priority order",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RulesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Targeting" ],
        "summary" : "Replace a flag's rules in an environment",
        "description" : "The whole ordered list at once, in one transaction, so no partial or reordered state is ever served (FR-RUL-004). A rule the engine could not apply is refused (FR-RUL-010).",
        "operationId" : "replaceRules",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReplaceRulesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "A rule cannot be applied; `errors` names it",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The rules as stored",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RulesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
    "/api/projects/{projectKey}/flags/{flagKey}/config/{envKey}/overrides" : {
      "get" : {
        "tags" : [ "Targeting" ],
        "summary" : "Get a flag's user overrides in an environment",
        "operationId" : "overrides",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The overrides",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OverridesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Targeting" ],
        "summary" : "Replace a flag's user overrides in an environment",
        "description" : "The whole set at once. Overrides apply to server keys only: a client key's ruleset carries none, because their user keys are real user identifiers (FR-KEY-008).",
        "operationId" : "replaceOverrides",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReplaceOverridesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The overrides as stored",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OverridesResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
    "/api/projects/{projectKey}/flags/{flagKey}/config/{envKey}" : {
      "get" : {
        "tags" : [ "Flags" ],
        "summary" : "Get a flag's configuration in an environment",
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The configuration",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagConfigResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Flags" ],
        "summary" : "Change a flag's configuration in an environment",
        "description" : "Kill switch, fallthrough value, rollout percentage and rollout salt. Takes effect on the environment's next ruleset. **Changing `rolloutSalt` re-buckets every user of this flag.**",
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "flagKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "projectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "envKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFlagConfigRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "The flag is archived",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "The configuration as changed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FlagConfigResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such project, environment or flag among yours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ReplaceRulesRequest" : {
        "type" : "object",
        "properties" : {
          "rules" : {
            "type" : "array",
            "description" : "In priority order. An empty list removes every rule (ADR-023: at most 100)",
            "items" : {
              "$ref" : "#/components/schemas/RuleRequest"
            },
            "maxItems" : 100,
            "minItems" : 0
          }
        },
        "required" : [ "rules" ]
      },
      "RuleRequest" : {
        "type" : "object",
        "properties" : {
          "attribute" : {
            "type" : "string",
            "example" : "country",
            "maxLength" : 100,
            "minLength" : 0
          },
          "operator" : {
            "type" : "string",
            "enum" : [ "EQUALS", "NOT_EQUALS", "IN", "NOT_IN", "CONTAINS", "STARTS_WITH", "ENDS_WITH" ],
            "example" : "IN",
            "minLength" : 1
          },
          "matchValues" : {
            "type" : "array",
            "description" : "Strings, numbers or booleans. Exactly one for EQUALS, NOT_EQUALS, CONTAINS, STARTS_WITH and ENDS_WITH, the last three strings only; one or more of one type for IN and NOT_IN (ADR-019)",
            "example" : [ "LK", "IN" ],
            "items" : { },
            "maxItems" : 1000,
            "minItems" : 1
          },
          "resultValue" : {
            "type" : "boolean",
            "description" : "What the flag evaluates to when the rule matches"
          }
        },
        "required" : [ "attribute", "matchValues", "operator", "resultValue" ]
      },
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "format" : "uri"
          },
          "properties" : {
            "type" : [ "object", "null" ],
            "additionalProperties" : { }
          }
        }
      },
      "RulesResponse" : {
        "type" : "object",
        "properties" : {
          "rules" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RuleRequest"
            }
          }
        },
        "required" : [ "rules" ]
      },
      "OverrideRequest" : {
        "type" : "object",
        "properties" : {
          "userKey" : {
            "type" : "string",
            "example" : "u-1042",
            "maxLength" : 256,
            "minLength" : 0
          },
          "value" : {
            "type" : "boolean"
          }
        },
        "required" : [ "userKey", "value" ]
      },
      "ReplaceOverridesRequest" : {
        "type" : "object",
        "properties" : {
          "overrides" : {
            "type" : "array",
            "description" : "An empty list removes every override (ADR-023: at most 1,000). Overrides do not apply to client keys, on either serving path (FR-KEY-008)",
            "items" : {
              "$ref" : "#/components/schemas/OverrideRequest"
            },
            "maxItems" : 1000,
            "minItems" : 0
          }
        },
        "required" : [ "overrides" ]
      },
      "OverridesResponse" : {
        "type" : "object",
        "properties" : {
          "overrides" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OverrideRequest"
            }
          }
        },
        "required" : [ "overrides" ]
      },
...
      "UpdateFlagConfigRequest" : {
        "type" : "object",
        "properties" : {
          "enabled" : {
            "type" : [ "boolean", "null" ],
            "description" : "The kill switch. Disabled returns false to everyone"
          },
          "fallthroughValue" : {
            "type" : [ "boolean", "null" ],
            "description" : "Returned when the flag is enabled and nothing else matched. While true, the rollout has no effect: everyone it does not reach gets true anyway"
          },
          "rolloutPercentage" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Whole percent of users, stored as basis points (ADR-011)",
            "maximum" : 100,
            "minimum" : 0
          },
          "rolloutSalt" : {
            "type" : [ "string", "null" ],
            "description" : "Changing the salt re-buckets every user of this flag: who has it at a given percentage changes. Give flags the same salt to roll them out to the same users (ADR-010)",
            "pattern" : "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
          }
        }
      },
      "FlagConfigResponse" : {
        "type" : "object",
        "properties" : {
          "flagKey" : {
            "type" : "string",
            "example" : "new-checkout"
          },
          "environment" : {
            "type" : "string",
            "example" : "production"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "offValue" : {
            "type" : "boolean",
            "description" : "Always false in v0.x (ADR-009)"
          },
          "fallthroughValue" : {
            "type" : "boolean"
          },
          "rolloutPercentage" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0
          },
          "rolloutSalt" : {
            "type" : "string",
            "example" : "new-checkout"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "enabled", "environment", "fallthroughValue", "flagKey", "offValue", "rolloutPercentage", "rolloutSalt", "updatedAt" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "An access token from `POST /api/auth/login`, valid for 8 hours.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "sdk-key" : {
        "type" : "http",
        "description" : "An API key issued for one environment. A server key reads every flag; a client key reads client-side-visible flags only, with no user overrides.",
        "scheme" : "bearer",
        "bearerFormat" : "flg_srv_… or flg_cli_…"
      }
    }
  }
}